added package.json
