+++ package.json
@@ -1,0 +1,26 @@
+{
+  "name": "pokedex-ci_api",
+  "version": "1.0.0",
+  "description": "Pokedex-CI_API",
+  "main": "index.js",
+  "dependencies": {
+    "cors": "^2.8.5",
+    "express": "^4.17.1",
+    "mongoose": "^5.6.7",
+    "unirest": "^0.6.0"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/team-red-cognz/Pokedex-CI_API.git"
+  },
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/team-red-cognz/Pokedex-CI_API/issues"
+  },
+  "homepage": "https://github.com/team-red-cognz/Pokedex-CI_API#readme"
+}
